+++ src/taxonomy.h
@@ -137,4 +137,4 @@ KHASH_MAP_INIT_STR(acc2asm, asm_tax);
 
 int parse_acc2tax(char* f, khash_t(acc2asm)*);
 
-int lca(int taxid0, int taxid1, taxonomy* tax);
+int lca(size_t taxid0, size_t taxid1, taxonomy* tax);
